requirements.txt
Django==3.1.5
gunicorn==20.0.4
python-dotenv==0.15.0
django-heroku==0.3.1
djangorestframework==3.12.2
Markdown==3.3.3
whitenoise==5.2.0
geopy==2.1.0